Migrate to a cache layer for authentication on the backend. Added Session persistence on the front

+++ loginservice/utils/utils.go
@@ -4,7 +4,7 @@ import (
 	"io"
 )
 
-func ParseBody(contentLength int64, body io.ReadCloser) []byte {
+func ParseBody(contentLength int, body io.ReadCloser) []byte {
 	content := make([]byte, contentLength)
 	body.Read(content)
 	return content
